+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - 2017-05-04
+### Added
+- Added support to get service return value in response middlewares.
+
 ## [1.0.1] - 2017-04-28
 ### Added
 - Version wildcards support for single '*' to match all.